updating angular version - started work on migrating cypress tests to latest version
