--- cypress.json
@@ -1,7 +1,0 @@
-{
-  "baseUrl": "http://127.0.0.1:4200",
-  "video": false,
-  "defaultCommandTimeout": 60000,
-  "chromeWebSecurity": false,
-  "retries": 2
-}
