+++ cypress/e2e/B_Data/analyses/analysis-detail.cy.js
