+++ cypress/e2e/B_Data/protocol/analysis/analysis-detail.cy.js
