+++ cypress/e2e/B_Data/publications/publication-detail.cy.js
