+++ cypress/e2e/B_Data/publications/publication.cy.js
