--- cypress/integration/C_Home/home.test.js
@@ -1,23 +1,0 @@
-import {HomePage} from "./home"
-
-describe('Home Page', () => {
-  beforeEach(() => {
-    cy.visit('/')
-  })
-
-  let homePage = new HomePage()
-
-  it('should display welcome message', () => {
-    homePage.check_page_contents()
-  });
-
-  it('should redirect to 404 when navigate to non-existing path', () => {
-    homePage.check_nonexistent_path()
-  });
-
-  it('should have 9 horizontal tabs', () => {
-    homePage.check_toolbar_menu()
-  });
-
-
-})
