+++ package.json
@@ -69,7 +69,7 @@
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "coveralls": "^3.0.9",
-    "cypress": "^9.6.1",
+    "cypress": "^13.6.1",
     "igv": "^2.15.11",
     "karma": "^6.3.20",
     "karma-chrome-launcher": "~3.1.0",
